+++ src/hin/conf.h
@@ -35,7 +35,7 @@
 #define HIN_HTTPD_ASYNC_STATX 0
 #define HIN_HTTPD_NULL_SERVER 0
 
-#define HIN_LINUX_BUG_5_11_3 0
+#define HIN_LINUX_BUG_5_11_3
 
 #endif