+++ src/hin/lua.h
@@ -21,6 +21,7 @@ typedef struct hin_server_data_struct {
   uint32_t debug;
   int cwd_fd;
   char * cwd_path;
+  void * cwd_dir;
   lua_State *L;
   struct hin_server_data_struct * next;
 } hin_server_data_t;
@@ -29,6 +30,8 @@ int lua_add_functions (lua_State * L, lua_function_t * func);
 
 int run_file (lua_State * L, const char * path);
 int run_function (lua_State * L, const char * name);
+
+int hin_server_set_work_dir (hin_server_data_t * server, const char * rel_path);
 
 #endif